feat(service): add user service - not completed
