+++ src/app/user/models/User.ts
@@ -3,3 +3,22 @@ export interface User {
   lastName: string;
   imageURL: string;
 }
+
+export interface LoginRequest {
+  username: string;
+  password: string;
+  rememberMe: boolean;
+}
+
+export interface LoginResponse {
+  firstName: string;
+  lastName: string;
+  imageURL: string | null;
+}
+
+export interface UserPermissions {
+  username: string;
+  firstName: string;
+  lastName: string;
+  permission: string;
+}
