--- src/app/user/services/auth/auth.model.ts
@@ -1,19 +1,0 @@
-// models.ts
-export interface LoginRequest {
-  username: string;
-  password: string;
-  rememberMe: boolean;
-}
-
-export interface LoginResponse {
-  firstName: string;
-  lastName: string;
-  imageURL: string | null;
-}
-
-export interface UserPermissions {
-  username: string;
-  firstName: string;
-  lastName: string;
-  permission: string;
-}
