+++ src/app/user/services/auth/auth.service.ts
@@ -48,8 +48,4 @@ export class AuthService {
         }),
       );
   }
-
-  getUserData() {
-    return this.userData$;
-  }
 }
